requirements.txt
openvino>=2024.0
openvino-genai>=2024.0
PyQt6>=6.5.0
PyYAML>=6.0
requests>=2.28.0
Pillow>=10.0.0
huggingface_hub>=0.20.0
numpy>=1.24.0
psutil>=5.9.0
watchdog>=3.0.0
html2text>=2020.1.16
beautifulsoup4>=4.12.0
jsonschema>=4.0.0
